backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
pandas==2.1.1
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.0
